Limit number of child processes
I used a semaphore with a initial positive value. Before forking, I wait
on the semaphore, and after getting SIGCHILD and waiting for a dead
child I post on the semaphore. The code seems to work as expected, but I
couldn't find a doc'tion that explains using semaphores in async code,
rather than a multi-thread code.